requirements.txt
APScheduler==3.10.4
Flask==3.0.3
lxml==5.3.0
pandas==2.2.3
python-dotenv==1.0.1
gunicorn==20.0.4
sqlalchemy==2.0.36